azure-function/requirements.txt
azure-functions
azure-cosmos>=4.5.0
azure-identity>=1.15.0
